extracted GasPriceFetcher
logic of using gasOracle or blockchain extracted
