+++ src/relayserver/RelayServer.ts
@@ -8,6 +8,7 @@ import { toBN, toHex } from 'web3-utils'
 import { IRelayHubInstance } from '../../types/truffle-contracts'
 
 import ContractInteractor, { TransactionRejectedByPaymaster } from '../relayclient/ContractInteractor'
+import { GasPriceFetcher } from '../relayclient/GasPriceFetcher'
 import { IntString } from '../relayclient/types/Aliases'
 import { RelayTransactionRequest, RelayTransactionRequestShape } from '../relayclient/types/RelayTransactionRequest'
 
@@ -51,6 +52,7 @@ export class RelayServer extends EventEmitter {
   alertedBlock: number = 0
   private initialized = false
   readonly contractInteractor: ContractInteractor
+  readonly gasPriceFetcher: GasPriceFetcher
   private readonly versionManager: VersionsManager
   private workerTask?: Timeout
   config: ServerConfigParams
@@ -74,6 +76,7 @@ export class RelayServer extends EventEmitter {
     this.versionManager = new VersionsManager(gsnRuntimeVersion)
     this.config = configureServer(config)
     this.contractInteractor = dependencies.contractInteractor
+    this.gasPriceFetcher = dependencies.gasPriceFetcher
     this.txStoreManager = dependencies.txStoreManager
     this.transactionManager = new TransactionManager(dependencies, this.config)
     this.managerAddress = this.transactionManager.managerKeyManager.getAddress(0)
@@ -498,7 +501,7 @@ latestBlock timestamp   | ${latestBlock.timestamp}
   }
 
   async _refreshGasPrice (): Promise<void> {
-    const gasPriceString = await this.contractInteractor.getGasPrice()
+    const gasPriceString = await this.gasPriceFetcher.getGasPrice()
     this.gasPrice = Math.floor(parseInt(gasPriceString) * this.config.gasPriceFactor)
     if (this.gasPrice === 0) {
       throw new Error('Could not get gasPrice from node')
