+++ test/relayserver/ServerTestEnvironment.ts
@@ -38,6 +38,7 @@ import PayMasterABI from '../../src/common/interfaces/IPaymaster.json'
 import { registerForwarderForGsn } from '../../src/common/EIP712/ForwarderUtil'
 import { RelayHubConfiguration } from '../../src/relayclient/types/RelayHubConfiguration'
 import { createServerLogger } from '../../src/relayserver/ServerWinstonLogger'
+import { GasPriceFetcher } from '../../src/relayclient/GasPriceFetcher'
 
 const Forwarder = artifacts.require('Forwarder')
 const StakeManager = artifacts.require('StakeManager')
@@ -172,8 +173,10 @@ export class ServerTestEnvironment {
     const managerKeyManager = this._createKeyManager(serverWorkdirs?.managerWorkdir)
     const workersKeyManager = this._createKeyManager(serverWorkdirs?.workersWorkdir)
     const txStoreManager = new TxStoreManager({ workdir: serverWorkdirs?.workdir ?? getTemporaryWorkdirs().workdir }, logger)
+    const gasPriceFetcher = new GasPriceFetcher('', '', this.contractInteractor, logger)
     const serverDependencies = {
       contractInteractor: this.contractInteractor,
+      gasPriceFetcher,
       logger,
       txStoreManager,
       managerKeyManager,
